# openapi.yaml
openapi: 3.0.3
info:
  version: 1.0.2
  title: E-library
  description: >-
    Endpoints:
    

    "/search" : query params: id(full match), (partial matching): title, author
    Search fields are a conjunction of partial matching.
    
    
    "/book" : search for a book using partial matching of: author, title. or full matching of book id


    "/borrow": All books in the borrow list must be validly available for borrowing (with valid stock). Otherwise the response will return a error.
    
    
    "/return": Similarly all books to be extended must be valid.
    
    
    "/extend": Extends books given email + loan ids
    
    
    Extra endpoints for flavour/ usability:
    
    
    "/add" : Books to be added to the library (for convenience of population. no testing)
    
    
    "/read": Returns the contents of the book
    
    
    "/loans": Returns the contents of the book
...
paths:
...
  /extend:
    summary: Extend a book loan
    post:
      summary: Extend book loans
      description: Extend books loan using loan Id and email (required). All must be valid, ie due date before Now(), loan id status must not be returned or previously extended already.
      operationId: extendLoan
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExtensionRequest'
      responses:
        '200':
          description: Success response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanReceipt'
        '204':
          description: No content found (no body)
        '400':
          description: Bad request or stock status invalid.
      tags:
        - loan
...
components:
  schemas:
...
    BookLoan:
      type: object
      properties:
        receipt_id:
          type: integer
          format: int64
          example: 1
        loan_id:
          type: integer
          format: int64
          example: 1
        book_id:
          type: integer
          format: int64
          example: 1
        start_date:
          type: integer
          format: int64
          example: 1741027052
        end_date:
          type: integer
          format: int64
          example: 1745260652
        status:
          type: integer
          format: int64
          example: 1
          description: >-
            Status of Bookloan - 1 = on_loan, 2 = returned, 3 = extended, 4 =
            expired (unused)
    LoanReceipt:
      type: object
      required:
        - email
        - receipt_id
      properties:
        receipt_id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Samuel
        email:
          type: string
          example: johndoe@mail.com
        loans:
          type: array
          items:
            $ref: '#/components/schemas/BookLoan'
        start_date:
          description: >-
            Ease of access field for Borrow receipts, will not be filled for return/extend. In epoch time
          type: integer
          format: int64
          example: 1741027052
        end_date:
          description: >-
            Ease of access field for Borrow receipts, will not be filled for return/extend. In epoch time
          type: integer
          format: int64
          example: 1745260652
...
    ExtensionRequest:
      type: object
      required:
        - loan_ids
        - email
      properties:
        loan_ids:
          type: array
          items:
            type: integer
        name:
          type: string
          example: Samuel
        email:
          type: string
          example: johndoe@mail.com
...